Fix vector search SQL type mismatch: cast input_divisor from text to bigint
The input_divisor column in sync_product_serverless_rates is text type, but
the CASE statement compared it against integer literals, causing psycopg2
UndefinedFunction error (operator does not exist: text = integer).

+++ backend/app/routes/reference/vector_search.py
@@ -64,9 +64,9 @@ def get_vector_search_modes(
             SELECT
                 cloud, size_or_model as mode, dbu_rate, input_divisor,
                 CASE
-                    WHEN input_divisor = 2000000 THEN 2
-                    WHEN input_divisor = 64000000 THEN 64
-                    ELSE input_divisor / 1000000.0
+                    WHEN input_divisor::bigint = 2000000 THEN 2
+                    WHEN input_divisor::bigint = 64000000 THEN 64
+                    ELSE input_divisor::bigint / 1000000.0
                 END as vector_capacity_millions
             FROM lakemeter.sync_product_serverless_rates
             {where_clause}
